SchemaLens-AI/backend/routes/connect.py
"""
routes/connect.py — DB connection simulation endpoint
"""

import asyncio
from fastapi import APIRouter
from fastapi.responses import StreamingResponse
from pydantic import BaseModel
from typing import Optional

router = APIRouter(prefix="/api/connect", tags=["connect"])


class ConnectRequest(BaseModel):
    db_type: str
    host: Optional[str] = "localhost"
    port: Optional[int] = 5432
    database: Optional[str] = ""
    schema_name: Optional[str] = "public"
    username: Optional[str] = ""


@router.post("/test")
async def test_connection(req: ConnectRequest):
    """Simulate a DB connection test."""
    steps = [
        {"text": f"→ Resolving hostname {req.host}...", "cls": "t-cyan"},
        {"text": f"→ TCP handshake on port {req.port}...", "cls": "t-cyan"},
        {"text": f"→ Authenticating as {req.username or 'user'}...", "cls": "t-cyan"},
        {"text": "→ Checking schema permissions...", "cls": "t-cyan"},
        {"text": "✓ Connection successful — latency: 12ms", "cls": "t-green"},
        {"text": f"✓ Found database: {req.database or 'main'}", "cls": "t-green"},
    ]
    return {"success": True, "steps": steps, "latency_ms": 12}


@router.post("/extract")
async def extract_schema(req: ConnectRequest):
    """Simulate async parallel schema extraction."""
    steps = [
        {"text": "→ Establishing connection...", "cls": "t-cyan"},
        {"text": f"→ Discovering schemas in {req.database or 'database'}...", "cls": "t-cyan"},
        {"text": "→ Async parallel schema extraction starting...", "cls": "t-muted"},
        {"text": "→ Extracting metadata [████░░░░] 12/34 tables...", "cls": "t-cyan"},
        {"text": "→ Extracting metadata [████████] 34/34 tables ✓", "cls": "t-cyan"},
        {"text": "→ Computing row counts, null rates, cardinality...", "cls": "t-cyan"},
        {"text": "→ Detecting PKs, FKs, indexes, constraints...", "cls": "t-cyan"},
        {"text": "→ Running AI Context Engine (GPT-4o)...", "cls": "t-muted"},
        {"text": "→ Generating business summaries for 34 tables...", "cls": "t-muted"},
        {"text": "→ Computing quality scores and anomaly baseline...", "cls": "t-muted"},
        {"text": "✓ Schema extraction complete! 34 tables · 287 columns · 100% AI-annotated", "cls": "t-green"},
        {"text": "✓ Database added to SchemaLens catalog.", "cls": "t-green"},
    ]
    return {"success": True, "steps": steps, "tables_extracted": 34, "columns_extracted": 287}
