SchemaLens-AI/backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
pandas==2.2.2
openai==1.30.1
python-multipart==0.0.9
scipy==1.13.0
numpy==1.26.4
python-dotenv==1.0.1
